Updating date and fixing bugs
removed brackets from date.
The bug happens when the function updateLastEditTime() is not called. The default text is "last edit111". Now the new default is blank, so that when the time cannot be updated or unavailable the text is left blank. The function only displays when the evaluation was created, not when the evaluation is last edited.